refactor, profile route added
