+++ backend/package.json
@@ -19,6 +19,7 @@
     "express-list-endpoints": "^7.1.0",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.4.0",
+    "multer": "^1.4.5-lts.1",
     "nodemon": "^3.0.1",
     "passport": "^0.7.0",
     "passport-local": "^1.0.0"
