+++ backend/server.js
@@ -2,9 +2,8 @@ import express from "express";
 import cors from "cors";
 import mongoose from "mongoose";
 import dotenv from "dotenv";
-import jwt from "jsonwebtoken";
+import authProfileRoutes from "./routes/authProfile";
 import passport from "./config/passport";
-import User from "./models/User";
 import expressListEndpoints from "express-list-endpoints";
 
 dotenv.config();
@@ -12,31 +11,6 @@ dotenv.config();
 const mongoUrl = process.env.MONGO_URL || "mongodb://localhost/project-final";
 mongoose.connect(mongoUrl);
 mongoose.Promise = global.Promise;
-
-// Function to generate JWT access token
-const generateAccessToken = (userId) => {
-  return jwt.sign({ userId }, process.env.JWT_SECRET, {
-    expiresIn: "24h",
-  });
-};
-
-// Middleware to athenticate the token
-const authenticateToken = async (req, res, next) => {
-  const authHeader = req.headers["authorization"];
-  const token = authHeader && authHeader.split(" ")[1];
-
-  if (token == null) {
-    return res.status(401).json({ error: "Token is missing" });
-  }
-
-  jwt.verify(token, process.env.JWT_SECRET, (err, user) => {
-    if (err) {
-      return res.status(403).json({ error: "Token is invalid" });
-    }
-    req.user = user;
-    next();
-  });
-};
 
 const port = process.env.PORT || 9000;
 const app = express();
@@ -46,49 +20,11 @@ app.use(cors());
 app.use(express.json());
 app.use(passport.initialize());
 
+app.use("/api", authProfileRoutes);
 // Start defining your routes here
 app.get("/", (req, res) => {
   const endpoints = expressListEndpoints(app);
   res.json(endpoints);
-});
-
-app.post("/users", async (req, res) => {
-  const { username, password, role } = req.body;
-  try {
-    const newUser = new User({ username, password, role });
-    await newUser.save();
-    const accessToken = generateAccessToken(newUser._id);
-    res.status(201).json({ id: newUser._id, accessToken });
-  } catch (error) {
-    console.error("Error registering user:", error);
-    if (error.code === 11000) {
-      res.status(400).json({ error: "Username already exists" });
-    } else {
-      res.status(500).json({ error: "Something went wrong" });
-    }
-  }
-});
-
-app.post("/sessions", (req, res, next) => {
-  passport.authenticate("local", { session: false }, (err, user, info) => {
-    if (err || !user) {
-      return res
-        .status(400)
-        .json({ error: info ? info.message : "Login failed" });
-    }
-    req.login(user, { session: false }, async (err) => {
-      if (err) {
-        return res.send(err);
-      }
-      const token = generateAccessToken(user._id);
-      return res.json({ user, token });
-    });
-  })(req, res, next);
-});
-
-// Authenticated endpoint
-app.get("/secrets", authenticateToken, (req, res) => {
-  res.json({ secret: "This is secret content" });
 });
 
 // Start the server
